--- a/app/__init__.py
+++ b/app/__init__.py
@@ -1,3 +1,3 @@
 def hello():
-    print("Hello world")
+    print("Hello World")