--- a/compose.yml
+++ b/compose.yml
@@ -4,3 +4,12 @@ services:
     tty: true
     working_dir: /app
     command: ["tail", "-f", "/dev/null"]
+    develop:
+      watch:
+        - action: sync
+          path: .
+          target: /app
+          ignore:
+            - .venv/
+        - action: rebuild
+          path: ./uv.lock