--- a/src/app/src/app/__init__.py
+++ b/src/app/src/app/__init__.py
@@ -1,2 +1,0 @@
-def hello() -> str:
-    return "Hello from app!"