--- a/src/app/src/app/main.py
+++ b/src/app/src/app/main.py
@@ -1,7 +1,0 @@
-def hello():
-    print("Hello, world!")
-
-
-if __name__ == "__main__":
-    hello()
-