## About

- enable byte code optimization
- enable cache system
- add .gitignore
- sync by `develop` (docker compose feature)

## Testing

build and run by docker command

```shell
uv_on_docker_playground (refactor/optimizaion) » docker build . -t uv_container
[+] Building 1.6s (10/10) FINISHED                         docker:desktop-linux
 => [internal] load build definition from Dockerfile                       0.0s
 => => transferring dockerfile: 578B                                       0.0s
 => [internal] load metadata for ghcr.io/astral-sh/uv:python3.12-bookworm  0.7s
 => [internal] load .dockerignore                                          0.0s
 => => transferring context: 2B                                            0.0s
 => [stage-0 1/5] FROM ghcr.io/astral-sh/uv:python3.12-bookworm-slim@sha2  0.0s
 => [internal] load build context                                          0.0s
 => => transferring context: 25.28kB                                       0.0s
 => CACHED [stage-0 2/5] WORKDIR /app                                      0.0s
 => [stage-0 3/5] RUN --mount=type=cache,target=/root/.cache/uv     --mou  0.3s
 => [stage-0 4/5] ADD . /app                                               0.0s
 => [stage-0 5/5] RUN --mount=type=cache,target=/root/.cache/uv     uv sy  0.5s
 => exporting to image                                                     0.0s
 => => exporting layers                                                    0.0s
 => => writing image sha256:578aa7b7323b7ba0a487b1f52c08cf610e39634b06a42  0.0s
 => => naming to docker.io/library/uv_container                            0.0s
uv_on_docker_playground (refactor/optimizaion) » docker run -it uv_container
Bytecode compiled 1 file in 46ms
Hello World
```

development with docker compose

```shell
uv_on_docker_playground (refactor/optimizaion) » docker compose watch
[+] Running 0/0
[+] Running 0/1 Building                                                   0.1s
[+] Building 0.6s (11/11) FINISHED                         docker:desktop-linux
 => [app internal] load build definition from Dockerfile                   0.0s
 => => transferring dockerfile: 578B                                       0.0s
 => [app internal] load metadata for ghcr.io/astral-sh/uv:python3.12-book  0.6s
 => [app internal] load .dockerignore                                      0.0s
 => => transferring context: 2B                                            0.0s
 => [app stage-0 1/5] FROM ghcr.io/astral-sh/uv:python3.12-bookworm-slim@  0.0s
 => [app internal] load build context                                      0.0s
 => => transferring context: 15.87kB                                       0.0s
 => CACHED [app stage-0 2/5] WORKDIR /app                                  0.0s
 => CACHED [app stage-0 3/5] RUN --mount=type=cache,target=/root/.cache/u  0.0s
 => CACHED [app stage-0 4/5] ADD . /app                                    0.0s
 => CACHED [app stage-0 5/5] RUN --mount=type=cache,target=/root/.cache/u  0.0s
 => [app] exporting to image                                               0.0s
 => => exporting layers                                                    0.0s
 => => writing image sha256:b7a11c1d3fa089bf9fee2855c97adc9a84f3d7811c021  0.0s
[+] Running 2/2o docker.io/library/uv_on_docker_playground-app             0.0s
 ✔ Service app                              Built                          0.8s
 ✔ Container uv_on_docker_playground-app-1  Started                       10.3s
Watch enabled
```

enter container and run

```shell
uv_on_docker_playground (refactor/optimizaion) » docker compose exec app bash
root@f76e7aa342d8:/app# uv run hello
Bytecode compiled 1 file in 24ms
Hello World
```

then edit `__init__.py`

```python
def hello():
    print("Hello World")
    print("edit")
```

now, the script outputs "edit" because of sync feature.

```shell
root@f76e7aa342d8:/app# uv run hello
Bytecode compiled 1 file in 33ms
Hello World
edit
```

when sync is detected, docker compose watch console show following message.

```
Syncing service "app" after 1 changes were detected
```